conver SuspenseErrorBoundary from javascript to typescript
